feat: 优化出价弹窗样式与统一弹窗主题
- 出价弹窗样式与UniApp保持一致(圆角10px、渐变按钮等)
- 添加DialogTheme统一定义弹窗样式
- 修复出价接口bidPrice参数类型(转Int)
- 新增AppBottomSheet通用底部弹窗组件
- 优化多处页面样式

+++ molitao_app/lib/data/repositories/auction_repository.dart
@@ -77,7 +77,7 @@ class AuctionRepository {
     try {
       final response = await _apiClient.dio.post(
         ApiEndpoints.bid,
-        data: {'auctionItemId': auctionItemId, 'bidPrice': bidPrice},
+        data: {'auctionItemId': auctionItemId, 'bidPrice': bidPrice.toInt()},
       );
 
       if (response.data != null) {
